1. code cleanup 2. creating APIs for task, solution
